Further research for #1118: default interface methods cause issues;
By running the tests with `-Xlog:redefine+class*=trace:stdout`, I gained
an insight into which method was actually causing the error. The minimal
testcase provided by sgerke-1L showed that this hunch was actually
correct, the DefaultMethodTest simulates the issue entirely without the
jOOQ dependency, showing that the default method is again the issue.